New classification algorithm and feature extraction method.
The new features is a different way of doing color histograms, in which
R, G, and B values are binned jointly.

The new algorithm finds the nearest point for each possible label, and
returns the nearest neighbor. It also returns a confidence based on how
close the test point is to the nearest neighbors.